another update of fixtures
